fix: removed active policy  data source
