--- docs/data-sources/active_policy.md
@@ -1,50 +1,0 @@
----
-# generated by https://github.com/hashicorp/terraform-plugin-docs
-page_title: "mondoo_active_policy Data Source - terraform-provider-mondoo"
-subcategory: ""
-description: |-
-  Data source that reports all active policies in a space
----
-
-# mondoo_active_policy (Data Source)
-
-Data source that reports all active policies in a space
-
-## Example Usage
-
-```terraform
-data "mondoo_active_policy" "policy" {
-  space_id = "your-space-1234567"
-}
-
-output "policies_mrn" {
-  value       = [for policy in data.mondoo_active_policy.policy.policies : policy.policy_mrn]
-  description = "The MRN of the policies in the space"
-}
-```
-
-<!-- schema generated by tfplugindocs -->
-## Schema
-
-### Optional
-
-- `space_id` (String) Space ID
-- `space_mrn` (String) Space MRN
-
-### Read-Only
-
-- `policies` (Attributes List) List of policies (see [below for nested schema](#nestedatt--policies))
-
-<a id="nestedatt--policies"></a>
-### Nested Schema for `policies`
-
-Read-Only:
-
-- `action` (String) Policies can be set to `Null`, `IGNORE` or `ACTIVE`
-- `assigned` (Boolean) Determines if a policy is enabled or disabled
-- `created_at` (String) Timestamp of policy creation
-- `is_public` (Boolean) Determines if a policy is public or private
-- `policy_mrn` (String) Unique policy Mondoo Resource Name
-- `policy_name` (String) Policy name
-- `updated_at` (String) Timestamp of policy update
-- `version` (String) Version of the policy
