--- examples/data-sources/mondoo_active_policy/main.tf
@@ -1,8 +1,0 @@
-terraform {
-  required_providers {
-    mondoo = {
-      source  = "mondoohq/mondoo"
-      version = ">= 0.4.0"
-    }
-  }
-}
